+++ libs/base_alloc/src/work/sync/reexports.rs
@@ -1,10 +1,6 @@
-// devela::work::sync::reexports
-//
-//!
+// devela_base_alloc::work::sync::reexports
 //
 // WAIT: [unique_rc_arc](https://github.com/rust-lang/rust/issues/112566)
-
-crate::mod_path!(std +pub _s "../../../libs/base_std/src/work/sync/reexports.rs");
 
 /* from either `alloc` or `portable-atomic-util` */
 
@@ -22,7 +18,7 @@ crate::items! {
     #[doc = "A thread-safe reference-counting pointer.\n\n"]
     #[cfg(all(feature = "alloc", not(feature = "dep_portable_atomic_util")))]
     #[cfg_attr(nightly_doc, doc(cfg(feature = "alloc")))]
-    pub use crate::_dep::_alloc::sync::Arc;
+    pub use alloc::sync::Arc;
 
     #[doc = crate::_TAG_ATOMIC!()]
     #[doc = crate::_TAG_ATOMIC_ALLOC_PORTABLE_UTIL!()]
@@ -36,5 +32,5 @@ crate::items! {
     #[doc = "A version of [`Arc`] that holds a non-owning reference.\n\n"]
     #[cfg(all(feature = "alloc", not(feature = "dep_portable_atomic_util")))]
     #[cfg_attr(nightly_doc, doc(cfg(feature = "alloc")))]
-    pub use crate::_dep::_alloc::sync::Weak as ArcWeak;
+    pub use alloc::sync::Weak as ArcWeak;
 }